comment-service

+++ core/comment-service/src/main/java/teamfive/service/CommentServiceImpl.java
@@ -1,23 +1,22 @@
-package teamfive.comment.service;
+package teamfive.service;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import teamfive.comment.dto.CommentDto;
-import teamfive.comment.dto.InputCommentDto;
-import teamfive.comment.dto.UpdateCommentDto;
-import teamfive.comment.mapper.CommentMapper;
-import teamfive.comment.model.Comment;
-import teamfive.comment.repository.CommentRepository;
-import teamfive.dto.event.EventResponseDto;
+import teamfive.dto.comment.CommentDto;
+import teamfive.dto.comment.InputCommentDto;
+import teamfive.dto.comment.UpdateCommentDto;
+import teamfive.dto.event.EventInternalDto;
 import teamfive.dto.user.UserDto;
 import teamfive.enums.EventState;
-import teamfive.event.model.Event;
-import teamfive.event.service.EventService;
 import teamfive.exception.ConflictException;
 import teamfive.exception.NotFoundException;
+import teamfive.feignclient.EventServiceClient;
 import teamfive.feignclient.UserServiceClient;
+import teamfive.mapper.CommentMapper;
+import teamfive.model.Comment;
+import teamfive.repository.CommentRepository;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -26,17 +25,18 @@ import java.util.List;
 @Service
 @AllArgsConstructor
 @Transactional(readOnly = true)
-public class CommentService {
+public class CommentServiceImpl implements CommentService {
     private final CommentRepository repository;
-    private final EventService eventService;
     private final UserServiceClient userClient;
     private final CommentMapper mapper;
+    private final EventServiceClient eventClient;
 
+    @Override
     @Transactional
     public CommentDto create(Long userId, InputCommentDto commentDto) {
         UserDto user = getUserDtoOrTrow(userId);
 
-        EventResponseDto event = eventService.getEventById(commentDto.getEventId());
+        EventInternalDto event = eventClient.getEventInternalById(commentDto.getEventId());
 
         validateEventState(event);
 
@@ -46,13 +46,17 @@ public class CommentService {
         return mapper.toCommentDto(repository.save(comment));
     }
 
+    @Override
     @Transactional
     public void deleteByIdByAdmin(Long id) {
-        checkExistsById(id);
+        if (!repository.existsById(id)) {
+            throw new NotFoundException("Комментарий с id={" + id + "} не найден");
+        }
         repository.deleteById(id);
         log.debug("Комментарий с id={} удален администратором", id);
     }
 
+    @Override
     @Transactional
     public void deleteForOwner(Long userId, Long commentId) {
         Comment comment = findCommentById(commentId);
@@ -65,68 +69,51 @@ public class CommentService {
         log.debug("Комментарий с id={} удален владельцем", commentId);
     }
 
-    public void checkExistsById(Long id) {
-        if (!repository.existsById(id)) {
-            throw new NotFoundException("Комментарий с id={" + id + "} не найден");
-        }
-    }
-
+    @Override
     public List<CommentDto> getByEventId(Long eventId) {
-        EventResponseDto eventResponseDto = eventService.getEventById(eventId);
-        return repository.getAllByEventId(eventId)
+        EventInternalDto eventDto = eventClient.getEventInternalById(eventId);
+        return repository.findAllByEventIdOrderByIdDesc(eventId)
                 .stream()
                 .map(mapper::toCommentDto)
                 .toList();
     }
 
+    @Override
     public List<CommentDto> getAllForUser(Long userId) {
         UserDto user = getUserDtoOrTrow(userId);
 
-        return repository.getAllByUserId(userId)
+        return repository.findAllByUserIdOrderByIdDesc(userId)
                 .stream()
                 .map(mapper::toCommentDto)
                 .toList();
     }
 
+    @Override
     @Transactional
     public CommentDto updateComment(Long commentId, Long userId, UpdateCommentDto dto) {
-        checkExistsById(commentId);
         Comment existingComment = findCommentById(commentId);
-
         validateCommentOwnership(userId, existingComment);
 
         if (dto.getText() != null) {
             existingComment.setText(dto.getText());
             repository.save(existingComment);
         }
-
-        return get(commentId);
+        return mapper.toCommentDto(existingComment);
     }
 
-    public CommentDto get(Long id) {
-        return repository.findById(id)
-                .map(mapper::toCommentDto)
-                .orElseThrow(() -> new NotFoundException("Комментарий с id={" + id + "} не найден"));
-    }
-
-    private void validateEventState(EventResponseDto event) {
-        if (!event.getState().equals(EventState.PUBLISHED.toString())) {
+    private void validateEventState(EventInternalDto event) {
+        if (event.getState() != EventState.PUBLISHED) {
             throw new ConflictException("Событие не опубликовано. Комментарии запрещены");
         }
     }
 
-    private Comment buildComment(UserDto user, EventResponseDto event, String text) {
-
-        Event relatedEvent = new Event();
-        relatedEvent.setId(event.getId());
-
-        Comment comment = new Comment();
-        comment.setUserId(user.getId());
-        comment.setEvent(relatedEvent);
-        comment.setText(text);
-        comment.setCreated(LocalDateTime.now());
-
-        return comment;
+    private Comment buildComment(UserDto user, EventInternalDto event, String text) {
+        return Comment.builder()
+                .userId(user.getId())
+                .eventId(event.getId())
+                .text(text)
+                .created(LocalDateTime.now())
+                .build();
     }
 
     private Comment findCommentById(Long commentId) {
